Refactor Doorkeeper auth and add config for apps
