+++ app/controllers/api/v1/order_articles_controller.rb
@@ -1,5 +1,5 @@
 class Api::V1::OrderArticlesController < Api::V1::BaseController
-  include CollectionScope
+  include Concerns::CollectionScope
 
   before_action :authenticate
