+++ app/controllers/api/v1/orders_controller.rb
@@ -1,5 +1,5 @@
 class Api::V1::OrdersController < Api::V1::BaseController
-  include CollectionScope
+  include Concerns::CollectionScope
 
   before_action :authenticate
