+++ plugins/discourse/lib/foodsoft_discourse/redirect_to_login.rb
@@ -4,13 +4,13 @@ module FoodsoftDiscourse
     def self.included(base) # :nodoc:
       base.class_eval do
 
-        alias orig_redirect_to_login redirect_to_login
+        alias foodsoft_discourse_orig_redirect_to_login redirect_to_login
 
         def redirect_to_login(options={})
           if FoodsoftDiscourse.enabled? && !FoodsoftConfig[:discourse_sso]
             redirect_to discourse_initiate_path
           else
-            orig_redirect_to_login(options)
+            foodsoft_discourse_orig_redirect_to_login(options)
           end
         end
 
@@ -22,5 +22,5 @@ end
 
 # modify existing helper
 ActiveSupport.on_load(:after_initialize) do
-  ApplicationController.send :include, FoodsoftDiscourse::RedirectToLogin
+  Concerns::Auth.send :include, FoodsoftDiscourse::RedirectToLogin
 end
